finished project

+++ users/api.py
@@ -1,41 +1,43 @@
 from django.contrib.auth.models import User
 from rest_framework.response import Response
-from rest_framework.views import APIView
 from users.serializers import UserSerializer
 from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.pagination import PageNumberPagination
+from users.permisions import UserPermission
+from rest_framework.viewsets import GenericViewSet
 
-class UserListAPI(APIView):
 
-    def get(self, request):
-        # Instanciar paginador
-        paginator = PageNumberPagination()
-        users = User.objects.all() # queryset
-        # paginar el queryset
-        paginator.paginate_queryset(users, request)
+class UserViewSet(GenericViewSet):
+
+    queryset = User
+    serializer_class = UserSerializer
+    pagination_class = PageNumberPagination
+    permission_classes = (UserPermission,)
+
+    def list(self, request):
+        users = User.objects.all()
+        self.paginate_queryset(users) # pagino el resultado
         serializer = UserSerializer(users, many=True)
-        serialized_users = serializer.data # lista de diccionarios
-        # devolver respueta paginada
-        return paginator.get_paginated_response(serialized_users)
+        return self.get_paginated_response(serializer.data) #devuelvo una respuesta paginada
     
-    def post(self, request):
+    def create(self, request):
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             new_user = serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-class UserDetailAPI(APIView):
 
-    def get(self, request, pk):
+    def retrieve(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
     
-    def put(self, request, pk):
+    def update(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(instance=user, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -43,7 +45,8 @@ class UserDetailAPI(APIView):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         
-    def delete(self, request, pk):
+    def destroy(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
